requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Django==4.2.14
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.4.1
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
twilio==9.3.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.11.0
